alevin_to_anndata.py: store as sparse matrix
For now, compute the dense version also (just in memory), to verify that the
results are identical.

Note that this is different than reading "dense" output from Salmon, as could
be specified with the `density` argument to this function. I had originally
adjusted this to convert to a SciPy CSR sparse matrix if density == 'sparse'
and a plain NumPy matrix if density == 'dense', then create the AnnData object
from whichever is appropriate, but I don't have any dense Salmon output handy
to test this with.

Rather than leave a completely untested code path, remove the
"density == 'dense'" branch entirely.